requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.3
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.86.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.9
requests==2.32.4
sniffio==1.3.1
starlette==0.40.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.23.2
